Code Refactor - Final version of lab 3

+++ Server-Side/Server.py
@@ -1,11 +1,9 @@
 import socket
-import json
-from Controller import Controller
-from ClientMongo import ClientMongo
+from Controller import *
 
 
 def server_program():
-    global conn
+    mongo = ClientMongo()
     try:
 
         host = socket.gethostname()
@@ -15,10 +13,8 @@ def server_program():
         server_socket.listen(2)
         conn, address = server_socket.accept()
         print("Connection from: " + str(address))
-        mongo = ClientMongo()
         mongo.update_mongoDB()
         data = ""
-
 
         while data != "exit":
 
@@ -30,7 +26,6 @@ def server_program():
                 command_type = commands[0].lower()
                 instance_type = commands[1].lower()
                 instance_name = commands[2].lower()
-                print(instance_name)
 
                 controller = Controller(command_type, instance_type, instance_name)
 
@@ -42,38 +37,38 @@ def server_program():
                     controller.drop_database()
                     mongo.update_mongoDB()
 
-                elif command_type == "create" and instance_type == "table":   # # create table table_name (int 1, FK int val2 ref table_name-table_column, varchar cc) on db_name
+                elif command_type == "create" and instance_type == "table":  # # create table table_name (int 1, FK int val2 ref table_name-table_column, varchar cc) on db_name
                     controller.create_table(client_request)
                     mongo.update_mongoDB()
 
-                elif command_type == "drop" and instance_type == "table":    # drop table table_name on database
-                    db_name = commands[4].upper()
-                    controller.delete_table(db_name)
+                elif command_type == "drop" and instance_type == "table":  # drop table table_name on database
+                    db_name = commands[4]
+                    if mongo.drop_table_mongoDB(db_name, instance_name):
+                        controller.delete_table(db_name)
                     mongo.update_mongoDB()
 
-                elif command_type == "create" and instance_type == "index":   # create  index unique/nonunique index_name on table_name (column_name,column_name) on database_name
+                elif command_type == "create" and instance_type == "index":  # create  index unique/non unique index_name on table_name (column_name,column_name) on database_name
                     table_name = commands[5].lower()
                     db_name = commands[len(commands) - 1].lower()
                     index_type = commands[2].lower()
                     index_name = commands[3].lower()
-                    controller.create_index(db_name, table_name,index_type,index_name,client_request)
+                    controller.create_index(db_name, table_name, index_type, index_name, client_request)
                     mongo.update_mongoDB()
 
-                elif command_type == "insert" and instance_type == "into":   # insert into db_name table_name values (1,2,3)
+                elif command_type == "insert" and instance_type == "into":  # insert into db_name table_name values (1,2,3)
                     table_name = commands[3].lower()
-                    id,attributes = controller.mongoDB_format(instance_name,table_name,client_request)
-                    mongo.add_mongoDB(id,attributes,instance_name,table_name)
+                    _id, attributes = controller.mongoDB_format(instance_name, table_name, client_request)
+                    mongo.insert_data_mongoDB(_id, attributes, instance_name, table_name)
                     mongo.update_mongoDB()
 
-                elif command_type == "delete" and instance_type == "from": #delete from db_name table_name value id_value
+                elif command_type == "delete" and instance_type == "from":  # delete from db_name table_name value id_value
                     table_name = commands[3].lower()
-                    id = commands[5].lower()
+                    _id = commands[5].lower()
                     try:  # convert the _id into an integer(if it was given as integer) or let it string
-                        id_value = int(id)
+                        id_value = int(_id)
                     except ValueError:
-                             id_value = id
-                    mongo.delete_mongoDB(id_value,instance_name,table_name)
-
+                        id_value = _id
+                    mongo.delete_data_mongoDB(id_value, instance_name, table_name)
 
                 print("\n>> Command executed")
 
@@ -90,7 +85,8 @@ def server_program():
 
     except Exception as e:
         print("An error occurred outside the loop:", str(e))
-    mongo.close_mongoDB()
+    finally:
+        mongo.close_mongoDB()
 
 
 if __name__ == "__main__":
